refactor!: move credential lookup methods from transaction context to client instance
